refactor(praxis-card): split PraxisCard.tsx into per-faction desktop modules (#839)
Pure move, no behaviour change. The nine desktop archetypes now live one per
module under components/praxisCard/desktop/, mirroring the existing
components/praxisCard/mobile/ layout, so each faction slice of the epic has a
disjoint footprint. components/PraxisCard.tsx keeps only the dispatcher and
re-exports ArchetypeProps.
