+++ frontend/src/factions/ua.ts
@@ -29,7 +29,7 @@ import UaProfileBody from '../pages/characterProfile/archetypes/UaProfileBody'
 import UaTaskCard from '../components/cards/UaTaskCard'
 import UaTaskDetail from '../pages/taskDetail/archetypes/UaTaskDetail'
 import UaVote from '../components/vote/UaVote'
-import { UaPraxisCard } from '../components/PraxisCard'
+import UaPraxisCard from '../components/praxisCard/desktop/UaPraxisCard'
 import { UaSigilAdapter } from '../components/cards/FactionSigil'
 import { UaCard } from '../components/cards/FactionCard'
 import { UaSelectCard } from '../components/cards/FactionSelectCard'
